+++ sdks/python/hatchet_sdk/__init__.py
@@ -3,7 +3,6 @@ from hatchet_sdk.clients.admin import (
     ScheduleTriggerWorkflowOptions,
     TriggerWorkflowOptions,
 )
-from hatchet_sdk.clients.dispatcher.action_listener import OTelAttribute
 from hatchet_sdk.clients.events import PushEventOptions
 from hatchet_sdk.clients.listeners.durable_event_listener import (
     RegisterDurableEventRequest,
@@ -150,6 +149,7 @@ from hatchet_sdk.runnables.types import (
     TaskDefaults,
     WorkflowConfig,
 )
+from hatchet_sdk.utils.opentelemetry import OTelAttribute
 from hatchet_sdk.waits import (
     Condition,
     OrGroup,
